+++ CHANGELOG.md
@@ -7,6 +7,10 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 ### Added
 
+## [Unreleased]
+### Added
+- csv and excel exports
+
 ## [102.0.3] - 2020-04-14
 ### Fixed
 - fixed issue with missing created_at column on fresh installation